+++ src/gr/di/netmanagement/Gps.java
@@ -58,8 +58,8 @@ public class Gps {
 		this.id = id;
 		this.email = email;
 		if (!(latitude.charAt(0) == 'N') && !(longtitude.charAt(0) == 'N')) {
-			this.location = new Location(Float.valueOf(latitude),
-					Float.valueOf(longtitude));
+			this.location = new Location(Double.valueOf(latitude),
+					Double.valueOf(longtitude));
 		} else {
 			this.location = new Location(-1.0f, -1.0f);
 		}
